+++ packages/upload-api/src/blob/accept.js
@@ -13,9 +13,6 @@ import {
 } from './lib.js'
 import * as HTTP from '@web3-storage/capabilities/http'
 
-const R2_REGION = 'auto'
-const R2_BUCKET = 'carpark-prod-0'
-
 /**
  * @param {API.W3ServiceContext} context
  * @returns {API.ServiceMethod<API.BlobAccept, API.BlobAcceptSuccess, API.BlobAcceptFailure>}
@@ -44,13 +41,12 @@ export function blobAcceptProvider(context) {
 
       const digest = Digest.decode(blob.digest)
       const content = createLink(rawCode, digest)
-      const url =
-        /** @type {API.URI<'https:'>} */
-        (
-          `https://w3s.link/ipfs/${content}?format=raw&origin=${encodeURIComponent(
-            `r2://${R2_REGION}/${R2_BUCKET}`
-          )}`
-        )
+      const createUrl = await context.blobsStorage.createDownloadUrl(
+        digest.bytes
+      )
+      if (createUrl.error) {
+        return createUrl
+      }
 
       const locationClaim = await Assert.location.delegate({
         issuer: context.id,
@@ -58,7 +54,7 @@ export function blobAcceptProvider(context) {
         with: context.id.toDIDKey(),
         nb: {
           content,
-          location: [url],
+          location: [createUrl.ok],
         },
         expiration: Infinity,
       })