+++ packages/upload-client/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [15.0.0](https://github.com/w3s-project/w3up/compare/upload-client-v14.1.1...upload-client-v15.0.0) (2024-05-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* delegated capabilities required to use `uploadFile`, `uploadDirectory` and `uploadCAR` have changed. In order to use these methods your agent will now need to be delegated `blob/add`, `index/add`, `filecoin/offer` and `upload/add` capabilities. Note: no code changes are required.
+
+### Features
+
+* generate sharded DAG index on client and invoke w `index/add` ([#1451](https://github.com/w3s-project/w3up/issues/1451)) ([a6d9026](https://github.com/w3s-project/w3up/commit/a6d9026536e60c0ce93b613acc6e337f2a21aeb2))
+
+## [14.1.1](https://github.com/w3s-project/w3up/compare/upload-client-v14.1.0...upload-client-v14.1.1) (2024-05-15)
+
+
+### Fixes
+
+* revert blob add by default ([#1456](https://github.com/w3s-project/w3up/issues/1456)) ([b77ec75](https://github.com/w3s-project/w3up/commit/b77ec750c3911f6f47fa32a5e1560858b347985c))
+* upload client export blob ([#1454](https://github.com/w3s-project/w3up/issues/1454)) ([5e55900](https://github.com/w3s-project/w3up/commit/5e55900870dcb72d7f53d0cd37990c8279b89b15))
+
 ## [14.1.0](https://github.com/w3s-project/w3up/compare/upload-client-v14.0.0...upload-client-v14.1.0) (2024-05-14)